+++ tests/test_kmers.py
@@ -4,6 +4,7 @@ multiprocessing.set_start_method('spawn', force=True)
 import unittest
 
 import numpy as np
+import pandas as pd
 from phylotypy import kmers
 
 
@@ -301,6 +302,29 @@ class TestGetKmers(unittest.TestCase):
         observed: list = kmers.genera_str_to_index(genera)
         expected: list = [0, 1, 2, 0, 1, 3]
         self.assertEqual(expected, observed)
+
+
+class TestSeqToKmersDatabase(unittest.TestCase):
+    def setUp(self) -> None:
+        self.kmer_size = 3
+        self.sequences = ["ATGCGCTA", "ATGCGCTC", "ATGCGCTC"]
+        self.genera = ["A", "B", "B"]
+        self.sequences_df = pd.DataFrame(dict(id=self.genera, sequence=self.sequences))
+
+    def test_make_kmer_database(self):
+        observed_idx, observed_kmers = kmers.seq_to_kmers_database(self.sequences_df, kmer_size=self.kmer_size)
+        expected_idx = [0, 1, 1]
+        self.assertEqual(observed_kmers[:, 0].tolist(), expected_idx)
+
+    def test_calc_priors_dense(self):
+        _, observed_kmers = kmers.seq_to_kmers_database(sequences_db=self.sequences_df, kmer_size=self.kmer_size)
+        observed_db = kmers.calc_priors_dense(np.array(observed_kmers), kmer_size=self.kmer_size)
+        self.assertEqual(observed_db[25], 0.875)
+
+    def test_conditional_prob(self):
+        observed_idx, observed_kmers = kmers.seq_to_kmers_database(self.sequences_df, kmer_size=self.kmer_size)
+        expected_idx = np.array([0, 1, 1])
+        self.assertTrue(np.array_equal(observed_kmers[:, 0], expected_idx))
 
 
 if __name__ == '__main__':